exception이 아닌 valid체크하여 로그인으로 던지기

+++ my-sup2/src/main/java/net/sup2/web/QuestionController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import net.sup2.domain.Question;
 import net.sup2.domain.QuestionRepository;
+import net.sup2.domain.Result;
 import net.sup2.domain.User;
 
 @Controller
@@ -57,51 +58,57 @@ public class QuestionController {
 	
 	@GetMapping("/{id}/form")
 	public String updateForm(@PathVariable Long id, Model model,HttpSession session) {
-		try {
+		
 			Question question = questionRepository.findById(id).get();
-			hasPermission(session, question);
+			Result result = valid(session, question);
+			if(!result.isValid()) {
+				model.addAttribute("errorMessage",result.getErrorMessage());
+				return "/user/login";
+			}
+							
 			model.addAttribute("question",question);
 			return "/qna/updateform";
-		} catch (IllegalStateException e) {
-			model.addAttribute("errorMessage",e.getMessage());
-			return "/user/login";
-		}		
+			
 	}
 	
-	private void hasPermission(HttpSession session, Question question) {
+	private Result valid(HttpSession session, Question question) {
 		if(!HttpSessionUtils.isLoginUser(session)) {
-			throw new IllegalStateException("로그인이 필요합니다.");
+			return Result.fail("로그인이 필요합니다.");
 		}
 		User loginUser = HttpSessionUtils.getUserFromSession(session);
 		if(!question.isSameUser(loginUser)) {
-			throw new IllegalStateException("자신이 쓴 글만 수정,삭제가 가능합니다.");
-		}		
-	}
+			return Result.fail("자신이 쓴 글만 수정,삭제가 가능합니다.");
+		}
+		
+		return Result.OK();		
+	}	
 	
 	@PutMapping("/{id}")
 	public String update(@PathVariable Long id,String title, String contents, HttpSession session, Model model) {
-		try {
-			Question question = questionRepository.findById(id).get();
-			hasPermission(session, question);
-			question.update(title, contents);
-			questionRepository.save(question);
-			return "redirect:/questions/{id}";
-		} catch (IllegalStateException e) {
-			model.addAttribute("message",e.getMessage());
+		
+		Question question = questionRepository.findById(id).get();
+		Result result = valid(session, question);
+		if(!result.isValid()) {
+			model.addAttribute("errorMessage",result.getErrorMessage());
 			return "/user/login";
-		}		
+		}
+		
+		question.update(title, contents);
+		questionRepository.save(question);
+		return "redirect:/questions/{id}";
+			
 	}
 	
 	@DeleteMapping("/{id}")
 	public String delete(@PathVariable Long id,HttpSession session,Model model) {
-		try {
-			Question question = questionRepository.findById(id).get();
-			hasPermission(session, question);
+		Question question = questionRepository.findById(id).get();
+		Result result = valid(session, question);
+		if(!result.isValid()) {
+			model.addAttribute("errorMessage",result.getErrorMessage());
+			return "/user/login";
+		}
+			
 			questionRepository.deleteById(id);			
-			return "redirect:/";
-		} catch (IllegalStateException e) {
-			model.addAttribute("message",e.getMessage());
-			return "/users/login";
-		}				
+			return "redirect:/";		 			
 	}	
 }
